move Footer and Navigation to App module
