+++ src/App/Footer/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Link from '../Link';
+import Link from '../../Link';
 
 import './style.css';
