+++ src/App/Navigation/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Link, withRouter } from 'react-router-dom';
 
-import * as routes from '../constants/routes';
-import Button from '../Button';
-import Input from '../Input';
+import * as routes from '../../constants/routes';
+import Button from '../../Button';
+import Input from '../../Input';
 
 import './style.css';
