+++ src/App/index.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
-import Navigation from '../Navigation';
-import Footer from '../Footer';
+import Navigation from './Navigation';
+import Footer from './Footer';
 import OrganizationPage from '../Organization';
 import ProfilePage from '../Profile';
